4.8 preview release merge
[CL 2568775 by Peter Knepley in UT-Clean branch]

+++ Engine/Plugins/Media/WmfMedia/Source/WmfMedia/Private/Tracks/WmfMediaVideoTrack.h
@@ -39,7 +39,7 @@ public:
 		}
 		else
 		{
-			FrameRate = Numerator / Denominator;
+			FrameRate = static_cast<float>(Numerator) / Denominator;
 		}
 	}
